Dashing fix
You can no longer dash spam if you are on the floor. You can dash multiple times in the air now if you are still in the air after the dash cooldown. The jump sound also now is fixed to only play with you can jump and not only when you press the jump button.